Updated Score
Created methods that the function score would call to help score the function.
Made an organizer method, multiple checker method, suit organizer method, and ranker method.
Ranker does the actual scoring once the hand has been organized into suits and multiples have been found.
Score method returns a double - the integer value (as if it were typecasted) is the score of the hand. After the decimal is the high card value.
e.g.
8.05 - Four of a Kind - Four 5
10 - Royal Flush
1.2 - No pair - High card 2 (should not be possible at all though)

The main method only produces a hypothetical situation for testing, but a switch statement is used to print the score of the hand.

Will still continue to revise when looking for score issues and bugs.
Example issue

	//NOTE FOR OTHER STUFF: Check rankings with two pair in the event both players have similar hands
	//Player1 - 5, 6
	//Player2 - 6, 3
	//Community - 6, 7, 7, 2, 9
	//Who wins? Tie?

UPDATE:
The tie issue has been dealt with by using kicker cards. If there is a tie among the kicker cards, then the pot is split.
Also added methods to display the ranking of one's hand and to determine the winner (or if there is an actual tie and the pot is split)
Still need to look into bugs and other possible issues.