Updated Score
Added more comments to clarify some points.
Possible bugs in determineWinnerVector and playerScorer. 

When needing to rank hands, call playerScorer. Pass the vector of players and the community hand. If player has folded, the player's score is 0. If the community hand is not complete yet (does not have five cards), the hand will be scored from the available cards. Any empty spots in the community hand, however, have to be defaulted to 0s (0 of spades. value = 0, suit = 's').

When needing to determine winner, call determineWinnerVector. Pass the vector of players and the community hand. determineWinnerVector could be more efficient, but a problem of converting Card to Card * was encountered when converting a vector to array.