Update Score
When needing to rank hands, call playerScorer. Pass the vector of players and the community hand. If player has folded, the player's score is 0. If the community hand is not complete yet (does not have five cards), the hand will be scored from the available cards. Any empty spots in the community hand, however, have to be defaulted to 0s (0 of spades. value = 0, suit = 's').

When needing to determine winner, call determineWinnerVector. Pass the vector of players and the community hand. determineWinnerVector could be more efficient, but a problem of converting Card to Card * was encountered when converting a vector to array.

Update: Added an additional if statement in determineWinnerVector -- forgot to account for a current score value that is higher than the current highValueScore.

Update2: Fixed bugs in determineWinnerVector - had to change increments of some loops, added an essential if statement, and fixed some unsigned int issues. Also added a hypothetical situation to the main function (Jack and John) to test out the determineWinnerVector and playerScorer.

Update3: Fixed bug with full house. Bug tested more vector examples.

Update4: Fixed bugs in the kicker from no pair to full house (whether it be bugs with ties or determining who the true winner is). If the kicker is called with a straight flush or royal flush, then it must be a tie.